Add random_cell method to Maze.
When I needed a random cell, the client would generate a random row
and column and call Maze::get. The upcoming work on masking, as well
as good design, makes it essential that the Maze itself find a
random cell so that it has total control over which cells are
available to be returned.