Show average longest paths for each algorithm.
I wanted to be able to see which algorithms tend to generate the
longest paths, in addition to being able to see average counts of
dead ends and whether the mazes are valid.

As part of this, I also changed the test functions to bring in the
maze sizes as parameters instead of using literals.